pyproject.toml
[project]
name = "18seconds"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.3.0",
    "reportlab>=4.5.0",
]
